requirements.txt
pytest
playwright
allure-pytest
openpyxl

